Added global float equality function
This will be useful in case there are any areas of code which compare float values. When I was initially programming the world generator I wasn't aware of the complications involving float equality tests so it's likely there are a few bugs in the program due to this.

Credit for this function goes to Grom on Stack Overflow:
https://stackoverflow.com/a/17467