added dear imgui to main loop along with a test gui
Added imgui to the program. Now the gui's for world gen settings and world gen need to be created. I'll likely ditch the separate menu/world gen/settings views and just have one single view with everything in it. This should be possible since it's easy to make compact. moveable, closeable windows with imgui.