switched noise-gen to FastNoise
Has more generation options than the old library, and is faster than libnoise, so it's a good improvement. The only downside is that the default world gen settings no longer work and need to be adjusted for some of the new settings. Will likely provide default values for each of the generation types at first.